Change all transaction float fields to integer and allow user to take query string for all or pagination

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -28,7 +28,7 @@ class CreateUsersTable extends Migration
                 $table->boolean('isActive')->index()->default(true);
                 $table->timestamp('email_verified_at')->nullable();
                 $table->string('password');
-                $table->float('earnings', 12, 2)->nullable();
+                $table->unsignedBigInteger('earnings')->nullable();
                 $table->string('avatar', 2048)->nullable();
                 $table->string('security_question')->nullable();
                 $table->longText('security_answer')->nullable();
